pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "german-ocr"
version = "0.6.0"
description = "High-performance German document OCR - Local & Cloud with GPU/CPU support"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Keyvan Hardani", email = "hello@keyvan.ai"}
]
maintainers = [
    {name = "Keyvan Hardani", email = "hello@keyvan.ai"}
]
keywords = [
    "ocr",
    "german",
    "document",
    "invoice",
    "text-extraction",
    "dsgvo",
    "llama-cpp",
    "gguf",
    "vision-language-model",
    "local-ai",
    "cpu-inference",
    "gpu-inference",
    "document-ai"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.9"
dependencies = [
    "Pillow>=9.0.0",
    "httpx>=0.24.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
hf = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "qwen-vl-utils>=0.0.10",
]
llamacpp = [
    "llama-cpp-python>=0.2.0",
]
cuda = [
    "torch>=2.0.0",
    "llama-cpp-python>=0.2.0",
]
all = [
    "german-ocr[hf,llamacpp]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
german-ocr = "german_ocr.cli:main"

[project.urls]
Homepage = "https://github.com/Keyvanhardani/german-ocr"
Documentation = "https://github.com/Keyvanhardani/german-ocr#readme"
Repository = "https://github.com/Keyvanhardani/german-ocr"
Issues = "https://github.com/Keyvanhardani/german-ocr/issues"
Changelog = "https://github.com/Keyvanhardani/german-ocr/releases"

[tool.setuptools.packages.find]
where = ["."]
include = ["german_ocr*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
